requirements.txt
Flask==2.3.3           # Основной фреймворк
Flask-SQLAlchemy==3.0.5 # ORM для работы с БД
Flask-Migrate==4.0.4   # Миграции базы данных
python-dotenv==1.0.0   # Загрузка переменных окружения
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Werkzeug==2.3.7
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0